requirements.txt
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
beautifulsoup4==4.12.2
blinker==1.7.0
bottle==0.12.25
bottle-websocket==0.2.9
browser-cookie3==0.19.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorzero==2.0
comtypes==1.2.0
cryptography==42.0.2
customtkinter==5.2.2
darkdetect==0.8.0
decorator==5.1.1
deep-translator==1.11.4
distro==1.9.0
dlib==19.24.2
dnspython==2.5.0
Eel==0.16.0
emoji==2.9.0
executing==2.0.1
Flask==3.0.0
Flask-Cors==4.0.0
future==0.18.3
gevent==23.9.1
gevent-websocket==0.10.1
google-api-core==2.15.0
google-auth==2.26.0
google-cloud-core==2.4.1
google-cloud-translate==3.13.0
googleapis-common-protos==1.62.0
gpiozero==2.0
greenlet==3.0.3
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.26.0
hyperframe==6.0.1
idna==3.6
ipython==8.17.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
langdetect==1.0.9
lz4==4.3.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
numpy==1.26.2
openai==1.6.1
opencv-python==4.9.0.80
packaging==23.2
parso==0.8.3
pefile==2023.2.7
pillow==10.2.0
prompt-toolkit==3.0.39
proto-plus==1.23.0
protobuf==4.25.1
psutil==5.9.7
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycparser==2.21
pycryptodomex==3.19.1
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
pygame==2.5.2
Pygments==2.16.1
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.11
pymongo==4.6.1
pyparsing==3.1.1
pypiwin32==223
pyttsx3==2.90
pywin32==306
pywin32-ctypes==0.2.2
requests==2.31.0
rsa==4.9
setuptools==69.0.3
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
soupsieve==2.5
SpeechRecognition==3.10.1
stack-data==0.6.3
tqdm==4.66.1
traitlets==5.13.0
transforms==0.2.1
typing_extensions==4.9.0
urllib3==2.1.0
waitress==2.1.2
wcwidth==0.2.9
Werkzeug==3.0.1
wheel==0.42.0
whichcraft==0.6.1
zope.event==5.0
zope.interface==6.1
